Conflicts:
	src/gui/config.txt
	src/lib/strategies/RatOne.cpp

+++ src/lib/solver/ThreadPool.h
@@ -86,7 +86,7 @@ namespace smtrat
             std::atomic_bool mDone;
             std::atomic_bool mPossibleOversubscription;
             unsigned mNumberOfCores;
-            unsigned mNumberOfThreads;
+            size_t mNumberOfThreads;
             unsigned mNumberOfRunningThreads;
             std::vector<std::thread*> mThreads;
             std::vector<std::condition_variable> mConditionVariables;
@@ -100,7 +100,7 @@ namespace smtrat
 
         public:
             // Constructor and destructor.
-            ThreadPool( unsigned, unsigned );
+            ThreadPool( size_t, unsigned );
             ~ThreadPool();
 
             // Public methods.
